User can choose what he want to do in bsaic calculator ('double' function to division)

+++ src/Calculator.java
@@ -14,7 +14,7 @@ public class Calculator {
     public int multiplication (int a, int b){
         return a*b;
     }
-    public int division (int a, int b){
+    public double division (double a, double b){
         return a/b;
     }
